added comments and moved code to /app/scripts for better seperation
